requirements.txt
Django>=4.2.0
tensorflow
Pillow>=10.0.0
numpy>=1.24.0
cryptography>=41.0.0
llama-cpp-python>=0.2.0


